==> scripts/00_libraries.R <==
# Load packages ---------------------------------------------------------------

library("here")
library("dplyr")
library("tidyr")
library("stringr")
library("glue")
library("forcats")
library("purrr")
library("ggplot2")
library("readr")
library("readxl")
library("fs")
library("brms")
library("emmeans")
library("ggdist")
library("tidybayes")
library("bayestestR")
library("patchwork")

# -----------------------------------------------------------------------------
